Data.Graph.Inductive.Internal.RootPath

Description

Inward directed trees as lists of paths.

Types

type RTree = [Path]

type LRTree a = [LPath a]

Operations

getPath :: Node -> RTree -> Path

getLPath :: Node -> LRTree a -> LPath a

getDistance :: Node -> LRTree a -> Maybe a

Return the distance to the given node in the given tree.

Returns Nothing if the given node is not reachable.

getLPathNodes :: Node -> LRTree a -> Path