Data.Graph.Inductive.Query.ArtPoint

Documentation

ap :: Graph gr => gr a b -> [Node]

Finds the articulation points for a connected undirected graph, by using the low numbers criteria:

a) The root node is an articulation point iff it has two or more children.

b) An non-root node v is an articulation point iff there exists at least one child w of v such that lowNumber(w) >= dfsNumber(v).