Data.Graph.Inductive.Query.BCC

Documentation

bcc :: DynGraph gr => gr a b -> [gr a b]

Finds the bi-connected components of an undirected connected graph. It first finds the articulation points of the graph. Then it disconnects the graph on each articulation point and computes the connected components.