Data.Graph.Inductive.Query.Indep

Description

Maximum Independent Node Sets

Documentation

indep :: DynGraph gr => gr a b -> [Node]

Calculate the maximum independent node set of the specified graph.

indepSize :: DynGraph gr => gr a b -> ([Node], Int)

The maximum independent node set along with its size.